Add ability for pbshm module sideloading
Framework comes packaged with a list of default
pbshm modules which have been tested and linked
into the framework; however, for testing and debugging
purposes, module authors need the ability to load
thier pbshm modules into the framework to ensure they
comply before being included in the framework.

The sideloader enables each pbshm module to encapsulate
the variables required for framework integration into a JSON
file. This JSON file can then be installed and uninstalled
from the framework to test the integration.