Register sideloaded modules into system
Calculate the site package location for each of the installed
sideloaded modules, import the module via the module_spec,
and register the imported module in the sytem modules
for the desired namespace. Lastly register the module blueprint
as a nested blueprint of the sideloader blueprint and add any
navigation entries into the app's config.