Override url_for method for nested blueprints
The sideloaded modules have no understanding that the module
is registered as a nested blueprint under the sideloader blueprint.
As such, the sideloaded modules will still reference any urls using
endpoints formatted as blueprint.method; however, the endpoints
are registered in the app as sideloader.blueprint.method.

The url_for method is therefore overridden in the context of the
sideloader blueprint, to prefix any requested endpoints for nested
blueprints with the sideloader prefix, before the request is sent to
the current app.