Add defaults for db root username and user email/password, and random db root password
Previously the root username and password for the MongoDB service and
the user password must be passed in using external secrets/environment
variables. This change makes these optional by adding default values for
the MongoDB root username and PBSHM Framework user email and password,
and autogenerating a random 32 digit hex string for the db root
password.

As a result, the MongoDB root user is set up during the initialisation
of the frontend container. This means that the MongoDB service used
during frontend initialisation must be a fresh MongoDB instance or
setting the db root user will fail.